requirements.txt
streamlit==0.64.0
pandas==1.1.0
requests==2.24.0
matplotlib==3.3.0
python-dotenv==0.14.0
